test_dependencyfile.py - suggested changes from deepsource
